feat(data-classes): Add S3ObjectSessionContext
